+++ app.py
@@ -276,9 +276,12 @@ def customer_portal(event, context):
             'body': json.dumps({"error": str(e)})
         }
 
-    del account['customer']
-    del account['subscription']
-    del account['subscription_item']
+    if 'customer' in account:
+        del account['customer']
+    if 'subscription' in account:
+        del account['subscription']
+    if 'subscription_item' in account:
+        del account['subscription_item']
 
     if session is None:
         account["portal_url"] = None