+++ app.py
@@ -2,7 +2,7 @@ from chalice import Chalice
 
 from chalicelib.telemetry import cloudwatch, xray_recorder
 from chalicelib.payments import customer_portal_url, customerportal_api_version
-from chalicelib.app_utils import process_request, validate_request_lambda
+from chalicelib.app_utils import process_request, validate_request_lambda, process_cors_preflight
 from chalicelib.auth import fetch_orgs, fetch_email_and_username, extract_client_version, userorganizations_api_version
 from chalicelib.log import mins_and_secs
 
@@ -204,6 +204,10 @@ def customer_portal(event, context):
 
     print(f'Inbound request {correlation_id} {function_name}')
 
+    preflight_response = process_cors_preflight(event)
+    if preflight_response:
+        return preflight_response
+
     try:
         # Extract parameters from the event object
         if 'body' in event:
@@ -318,6 +322,10 @@ def user_organizations(event, context):
     function_name = context.function_name
 
     print(f'Inbound request {correlation_id} {function_name}')
+
+    preflight_response = process_cors_preflight(event)
+    if preflight_response:
+        return preflight_response
 
     try:
         # Extract parameters from the event object